Sharp js installed to resize images
End user can now just submit any type of picture; We will handle that it is appropriate and save to server with Multer. With that image, we then use Sharp js to resize it appropriately. This has significantly improved load times on the client end.

Original Multer file should be deleted once done.